% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{print.gofgamma}
\alias{print.gofgamma}
\title{Print method for tests of Gamma distribution}
\usage{
\method{print}{gofgamma}(x, ...)
}
\arguments{
\item{x}{object of class "gofgamma".}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
the argument x, invisibly, as for all \link{print} methods.
}
\description{
Printing objects of class "gofgamma".
}
\details{
A \code{gofgamma} object is a named list of numbers and character string, supplemented with \code{test} (the name of the teststatistic). \code{test} is displayed as a title.
The remaining elements are given in an aligned "name = value" format.
}
\examples{
\donttest{print(test.BE(rgamma(20,1)))}

}
